// thunder-sdk/src/types/transfer_managers/transfer_manager_721/factories/TransferManager721Abi__factory.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */

import type { Provider, BaseWalletLocked, AbstractAddress } from "fuels";
import { Interface, Contract } from "fuels";
import type {
  TransferManager721Abi,
  TransferManager721AbiInterface,
} from "../TransferManager721Abi";
const _abi = {
  types: [
    {
      typeId: 0,
      type: "()",
      components: [],
      typeParameters: null,
    },
    {
      typeId: 1,
      type: "b256",
      components: null,
      typeParameters: null,
    },
    {
      typeId: 2,
      type: "enum Identity",
      components: [
        {
          name: "Address",
          type: 5,
          typeArguments: null,
        },
        {
          name: "ContractId",
          type: 6,
          typeArguments: null,
        },
      ],
      typeParameters: null,
    },
    {
      typeId: 3,
      type: "str[21]",
      components: null,
      typeParameters: null,
    },
    {
      typeId: 4,
      type: "str[36]",
      components: null,
      typeParameters: null,
    },
    {
      typeId: 5,
      type: "struct Address",
      components: [
        {
          name: "value",
          type: 1,
          typeArguments: null,
        },
      ],
      typeParameters: null,
    },
    {
      typeId: 6,
      type: "struct ContractId",
      components: [
        {
          name: "value",
          type: 1,
          typeArguments: null,
        },
      ],
      typeParameters: null,
    },
    {
      typeId: 7,
      type: "u64",
      components: null,
      typeParameters: null,
    },
  ],
  functions: [
    {
      inputs: [],
      name: "get_exchange",
      output: {
        name: "",
        type: 6,
        typeArguments: null,
      },
    },
    {
      inputs: [
        {
          name: "exchange_contract",
          type: 6,
          typeArguments: null,
        },
      ],
      name: "initialize",
      output: {
        name: "",
        type: 0,
        typeArguments: null,
      },
    },
    {
      inputs: [
        {
          name: "collection",
          type: 6,
          typeArguments: null,
        },
        {
          name: "from",
          type: 2,
          typeArguments: null,
        },
        {
          name: "to",
          type: 2,
          typeArguments: null,
        },
        {
          name: "token_id",
          type: 7,
          typeArguments: null,
        },
        {
          name: "amount",
          type: 7,
          typeArguments: null,
        },
      ],
      name: "transfer_nft",
      output: {
        name: "",
        type: 0,
        typeArguments: null,
      },
    },
  ],
  loggedTypes: [
    {
      logId: 0,
      loggedType: {
        name: "",
        type: 3,
        typeArguments: null,
      },
    },
    {
      logId: 1,
      loggedType: {
        name: "",
        type: 4,
        typeArguments: null,
      },
    },
  ],
};

export class TransferManager721Abi__factory {
  static readonly abi = _abi;
  static createInterface(): TransferManager721AbiInterface {
    return new Interface(_abi) as unknown as TransferManager721AbiInterface;
  }
  static connect(
    id: string | AbstractAddress,
    walletOrProvider: BaseWalletLocked | Provider
  ): TransferManager721Abi {
    return new Contract(
      id,
      _abi,
      walletOrProvider
    ) as unknown as TransferManager721Abi;
  }
}